Fix phase AD minor cleanup findings
