+++ crates/atm-core/src/lib.rs
@@ -8,6 +8,8 @@ pub mod boundary;
 /// adapter crates.
 #[doc(hidden)]
 pub mod boundary_support;
+/// Shared caller-context resolution and ATM-owned environment parsing helpers.
+pub mod caller_context;
 /// Mailbox cleanup workflows for read and acknowledged messages.
 pub mod clear;
 /// Internal configuration discovery and resolution helpers.
