+++ docs/plans/phase-AD/plan-phase-AD.md
@@ -1,6 +1,6 @@
 ---
 title: Phase AD Plan
-status: active
+status: complete
 branch: integrate/phase-AD
 worktree: /Users/randlee/Documents/github/atm-core-worktrees/integrate/phase-AD
 ---
